brake superclass into multiple classes
