+++ .gitignore
@@ -1,0 +1,3 @@
+__pycache__/*
+config.ini
+test_lib.py
